+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Hydrophone is the module responsible for sending emails.
 This API sends notifications to users for things like forgotten passwords, initial signup, and invitations.
 
+## unreleased
+### Fixed 
+- YLP-531 Reset key can be null on a reset password api call
 ## 1.6.0 - 2021-04-28
 ### Changed
 - YLP-559 emails which are part of URLs are URL encoded